+++ components/bridge/dao/impl/src/test/java/com/flipperdevices/bridge/dao/impl/api/key/FileComparatorTest.kt
@@ -1,59 +1,47 @@
 package com.flipperdevices.bridge.dao.impl.api.key
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.flipperdevices.bridge.dao.impl.comparator.DefaultFileComparator
 import com.flipperdevices.bridge.dao.impl.comparator.FileComparator
-import com.flipperdevices.bridge.dao.impl.comparator.FileComparatorExt.isSameContent
+import com.flipperdevices.core.FlipperStorageProvider
+import com.flipperdevices.core.sourceOrEmpty
+import com.flipperdevices.core.test.FlipperStorageProviderTestRule
 import kotlinx.coroutines.test.runTest
-import org.junit.After
+import okio.Path.Companion.toPath
+import okio.buffer
+import okio.use
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.TemporaryFolder
 import org.junit.runner.RunWith
-import java.io.File
-import java.io.InputStream
 
 @RunWith(AndroidJUnit4::class)
 class FileComparatorTest {
     private lateinit var fileComparator: FileComparator
 
     @get:Rule
-    val folder: TemporaryFolder = TemporaryFolder()
-    private lateinit var tempFolder: File
+    val storageProviderRule = FlipperStorageProviderTestRule()
+    private lateinit var storageProvider: FlipperStorageProvider
 
     @Before
     fun setUp() {
-        tempFolder = folder.newFolder()
-        fileComparator = DefaultFileComparator()
-    }
-
-    @After
-    fun onStop() {
-        tempFolder.delete()
-    }
-
-    /**
-     * Create empty [InputStream] for non-existing files
-     */
-    private fun File.inputStreamOrEmpty(): InputStream {
-        return if (exists()) {
-            inputStream()
-        } else {
-            InputStream.nullInputStream()
-        }
+        storageProvider = storageProviderRule.flipperStorageProvider
+        fileComparator = FileComparator(storageProvider)
     }
 
     @Test
     fun GIVEN_not_existing_file_WHEN_is_equal_THEN_true() {
         runTest {
-            val file1 = File("FILE_NAME_1")
-            val file2 = File("FILE_NAME_2")
-            fileComparator.isSameContent(
-                file1.inputStreamOrEmpty(),
-                file2.inputStreamOrEmpty(),
-            ).run(Assert::assertTrue)
+            val file1 = "FILE_NAME_1".toPath()
+            val file2 = "FILE_NAME_2".toPath()
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source1,
+                        source2
+                    ).run(Assert::assertTrue)
+                }
+            }
             fileComparator.isSameContent(
                 file1,
                 file2,
@@ -64,14 +52,20 @@ class FileComparatorTest {
     @Test
     fun GIVEN_not_existing_and_existing_file_WHEN_is_equal_THEN_false() {
         runTest {
-            val file1 = File(tempFolder, "FILE_1").apply {
-                writeText("RANDOM_CONTENT")
+            val file1 = "FILE_1".toPath()
+            storageProvider.fileSystem.write(file1) {
+                writeUtf8("RANDOM_CONTENT")
             }
-            val file2 = File("ANOTHER_RANDOM_NAME")
-            fileComparator.isSameContent(
-                file1.inputStreamOrEmpty(),
-                file2.inputStreamOrEmpty(),
-            ).run(Assert::assertFalse)
+            val file2 = "ANOTHER_RANDOM_NAME".toPath()
+
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source1,
+                        source2
+                    ).run(Assert::assertFalse)
+                }
+            }
             fileComparator.isSameContent(
                 file1,
                 file2,
@@ -82,16 +76,22 @@ class FileComparatorTest {
     @Test
     fun GIVEN_two_same_content_files_WHEN_is_equal_THEN_true() {
         runTest {
-            val file1 = File(tempFolder, "FILE_1").apply {
-                writeText("STUB_CONTENT")
+            val file1 = "FILE_1".toPath()
+            storageProvider.fileSystem.write(file1) {
+                writeUtf8("STUB_CONTENT")
             }
-            val file2 = File(tempFolder, "FILE_2").apply {
-                writeText("STUB_CONTENT")
+            val file2 = "FILE_2".toPath()
+            storageProvider.fileSystem.write(file2) {
+                writeUtf8("STUB_CONTENT")
             }
-            fileComparator.isSameContent(
-                file1.inputStreamOrEmpty(),
-                file2.inputStreamOrEmpty(),
-            ).run(Assert::assertTrue)
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source1,
+                        source2
+                    ).run(Assert::assertTrue)
+                }
+            }
             fileComparator.isSameContent(
                 file1,
                 file2,
@@ -102,16 +102,56 @@ class FileComparatorTest {
     @Test
     fun GIVEN_two_different_content_files_WHEN_is_equal_THEN_false() {
         runTest {
-            val file1 = File(tempFolder, "FILE_1").apply {
-                writeText("FILLED_CONTENT_1")
+            val file1 = "FILE_1".toPath()
+            storageProvider.fileSystem.write(file1) {
+                writeUtf8("FILLED_CONTENT_1")
             }
-            val file2 = File(tempFolder, "FILE_2").apply {
-                writeText("FILLED_CONTENT_2")
+            val file2 = "FILE_2".toPath()
+            storageProvider.fileSystem.write(file2) {
+                writeUtf8("FILLED_CONTENT_2")
+            }
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source1,
+                        source2
+                    ).run(Assert::assertFalse)
+                }
             }
             fileComparator.isSameContent(
-                file1.inputStreamOrEmpty(),
-                file2.inputStreamOrEmpty(),
+                file1,
+                file2
             ).run(Assert::assertFalse)
+        }
+    }
+
+    @Test
+    fun GIVEN_two_different_size_files_WHEN_is_equal_THEN_false() {
+        runTest {
+            val file1 = "FILE_1".toPath()
+            storageProvider.fileSystem.write(file1) {
+                writeUtf8("FILLED_CONTENT_1")
+            }
+            val file2 = "FILE_2".toPath()
+            storageProvider.fileSystem.write(file2) {
+                writeUtf8("FILLED_CONTENT")
+            }
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source1,
+                        source2
+                    ).run(Assert::assertFalse)
+                }
+            }
+            storageProvider.fileSystem.sourceOrEmpty(file1).buffer().use { source1 ->
+                storageProvider.fileSystem.sourceOrEmpty(file2).buffer().use { source2 ->
+                    fileComparator.isSameContent(
+                        source2,
+                        source1
+                    ).run(Assert::assertFalse)
+                }
+            }
             fileComparator.isSameContent(
                 file1,
                 file2