+++ components/core/atomicfile/src/androidMain/kotlin/com/flipperdevices/core/atomicfile/SinkWithOutputStream.kt
@@ -1,0 +1,12 @@
+package com.flipperdevices.core.atomicfile
+
+import okio.Sink
+import okio.sink
+import java.io.FileOutputStream
+
+actual class SinkWithOutputStream private constructor(
+    private val sink: Sink,
+    val outputStream: FileOutputStream
+) : Sink by sink {
+    constructor(outputStream: FileOutputStream) : this(outputStream.sink(), outputStream)
+}