+++ components/core/atomicfile/src/desktopMain/kotlin/com/flipperdevices/core/atomicfile/SinkWithOutputStream.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.core.atomicfile
+
+import okio.Sink
+
+actual class SinkWithOutputStream(sink: Sink) : Sink by sink
+
+fun Sink.wrap(): SinkWithOutputStream = SinkWithOutputStream(this)