+++ components/nfceditor/impl/src/main/java/com/flipperdevices/nfceditor/impl/viewmodel/NfcEditorViewModel.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.nfceditor.impl.viewmodel
 
-import android.app.Application
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
 import com.flipperdevices.bridge.dao.api.delegates.key.UpdateKeyApi
 import com.flipperdevices.bridge.dao.api.model.FlipperKey
@@ -10,7 +9,7 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.ui.hexkeyboard.HexKey
 import com.flipperdevices.core.ui.lifecycle.DecomposeViewModel
 import com.flipperdevices.keyedit.api.NotSavedFlipperKey
-import com.flipperdevices.keyedit.api.toNotSavedFlipperFile
+import com.flipperdevices.keyedit.api.NotSavedFlipperKeyApi
 import com.flipperdevices.keyparser.api.KeyParser
 import com.flipperdevices.keyparser.api.model.FlipperKeyParsed
 import com.flipperdevices.metric.api.MetricApi
@@ -29,12 +28,12 @@ import kotlinx.coroutines.launch
 @Suppress("LongParameterList")
 class NfcEditorViewModel @AssistedInject constructor(
     @Assisted private val flipperKeyPath: FlipperKeyPath,
-    private val application: Application,
     private val keyParser: KeyParser,
     private val updateKeyApi: UpdateKeyApi,
     private val synchronizationApi: SynchronizationApi,
     private val simpleKeyApi: SimpleKeyApi,
-    private val metricApi: MetricApi
+    private val metricApi: MetricApi,
+    private val notSavedFlipperKeyApi: NotSavedFlipperKeyApi
 ) : DecomposeViewModel(), LogTagProvider {
     override val TAG = "NfcEditorViewModel"
 
@@ -122,7 +121,7 @@ class NfcEditorViewModel @AssistedInject constructor(
                 localNfcEditorState
             )
             val notSavedKey = NotSavedFlipperKey(
-                mainFile = newFlipperKey.mainFile.toNotSavedFlipperFile(application),
+                mainFile = notSavedFlipperKeyApi.toNotSavedFlipperFile(newFlipperKey.mainFile),
                 additionalFiles = listOf(),
                 notes = newFlipperKey.notes
             )