+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/presentation/viewmodel/SaveTempSignalViewModel.kt
@@ -1,12 +1,11 @@
 package com.flipperdevices.remotecontrols.impl.setup.presentation.viewmodel
 
-import android.content.Context
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
 import com.flipperdevices.bridge.rpc.api.FlipperStorageApi
+import com.flipperdevices.core.FlipperStorageProvider
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ktx.jre.launchWithLock
 import com.flipperdevices.core.log.LogTagProvider
-import com.flipperdevices.core.preference.FlipperStorageProvider
 import com.flipperdevices.core.ui.lifecycle.DecomposeViewModel
 import com.flipperdevices.remotecontrols.api.SaveTempSignalApi
 import com.squareup.anvil.annotations.ContributesBinding
@@ -21,7 +20,7 @@ private const val EXT_PATH = "/ext"
 @ContributesBinding(AppGraph::class, SaveTempSignalApi::class)
 class SaveTempSignalViewModel @Inject constructor(
     private val flipperStorageApi: FlipperStorageApi,
-    private val context: Context
+    private val storageProvider: FlipperStorageProvider
 ) : DecomposeViewModel(),
     LogTagProvider,
     SaveTempSignalApi {
@@ -38,8 +37,8 @@ class SaveTempSignalViewModel @Inject constructor(
             _state.emit(SaveTempSignalApi.State.Uploading(0f))
             launchWithLock(mutex, viewModelScope, "load") {
                 filesDesc.forEachIndexed { index, fileDesc ->
-                    FlipperStorageProvider.useTemporaryFile(context) { deviceFile ->
-                        deviceFile.writeText(fileDesc.textContent)
+                    storageProvider.useTemporaryFile { deviceFile ->
+                        deviceFile.toFile().writeText(fileDesc.textContent)
                         val fAbsolutePath = FlipperFilePath(
                             folder = fileDesc.extFolderPath,
                             nameWithExtension = fileDesc.nameWithExtension
@@ -47,7 +46,7 @@ class SaveTempSignalViewModel @Inject constructor(
                         flipperStorageApi.mkdirs("$EXT_PATH/${fileDesc.extFolderPath}")
                         flipperStorageApi.upload(
                             pathOnFlipper = fAbsolutePath,
-                            fileOnAndroid = deviceFile,
+                            fileOnAndroid = deviceFile.toFile(),
                             progressListener = { currentProgress ->
                                 _state.value = SaveTempSignalApi.State.Uploading(
                                     progressPercent = index + currentProgress